+++ apps/studio/components/layouts/ProjectLayout/LayoutHeader/HelpDropdown/HelpDropdown.constants.ts
@@ -1,5 +1,3 @@
-import type { SupportFormUrlKeys } from 'components/interfaces/Support/SupportForm.utils'
-
 export type HelpOptionId =
   | 'assistant'
   | 'docs'
@@ -28,15 +26,4 @@ export const ASSISTANT_SUGGESTIONS = {
       { label: 'RLS Setup', description: 'Implement row level security for my tables' },
     ],
   },
-}
-
-export function getSupportLinkQueryParams(
-  project: { parent_project_ref?: string } | undefined,
-  org: { slug?: string } | undefined,
-  routerRef: string | undefined
-): Partial<SupportFormUrlKeys> | undefined {
-  const projectRef = project?.parent_project_ref ?? routerRef
-  if (projectRef) return { projectRef }
-  if (org?.slug) return { orgSlug: org.slug }
-  return undefined
 }
