--- apps/studio/components/layouts/ProjectLayout/LayoutHeader/HelpOptions/index.ts
@@ -1,8 +1,0 @@
-export {
-  ASSISTANT_SUGGESTIONS,
-  getSupportLinkQueryParams,
-  HELP_OPTION_IDS,
-} from './helpOptionsConfig'
-export type { HelpOptionId } from './helpOptionsConfig'
-export { HelpSection } from './HelpSection'
-export { HelpOptionsList } from './HelpOptionsList'
