+++ apps/studio/components/layouts/ProjectLayout/LayoutHeader/MergeRequestButton.tsx
@@ -1,7 +1,3 @@
-import { GitMerge } from 'lucide-react'
-import { useRouter } from 'next/router'
-import { toast } from 'sonner'
-
 import { useParams } from 'common'
 import { ButtonTooltip } from 'components/ui/ButtonTooltip'
 import { useBranchUpdateMutation } from 'data/branches/branch-update-mutation'
@@ -9,6 +5,9 @@ import { useBranchesQuery } from 'data/branches/branches-query'
 import { useSendEventMutation } from 'data/telemetry/send-event-mutation'
 import { useSelectedOrganizationQuery } from 'hooks/misc/useSelectedOrganization'
 import { useSelectedProjectQuery } from 'hooks/misc/useSelectedProject'
+import { GitMerge } from 'lucide-react'
+import { useRouter } from 'next/router'
+import { toast } from 'sonner'
 
 export const MergeRequestButton = () => {
   const { ref } = useParams()
@@ -91,5 +90,3 @@ export const MergeRequestButton = () => {
     />
   )
 }
-
-export default MergeRequestButton
